feat: add get-source-incomplete-entries helper script
The idea of this script is to identify entries in the database that
are inconsistent with regards to the source letters they have.

In this context, an inconsistent character X would be a character that is used as a component
in the IDS of a second character Y with source region letters ABCD, where X does not have
all ABCD in its own IDS options.

For instance, the character 㐊 entry read as:
U+340A  㐊      ⿱士乙(K)
However, 㐊 is used in the IDS of 𥑄, which reads as:
U+25444 𥑄      ⿰石㐊(GT)
Such IDS is defined for G and T, making 㐊 inconsistent.